requirements.txt
appnope==0.1.2
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-cors-headers==3.1.1
django-countries==6.1.3
django-environ==0.4.5
django-filer==1.7.1
django-import-export==2.3.0
django-jinja==2.6.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-middleware-global-request==0.2.0
django-model-utils==4.0.0
django-mptt==0.11.0
django-polymodels==1.6
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.9
django-utils-six==2.0
djangorestframework==3.11.1
easy-thumbnails==2.7
et-xmlfile==1.0.1
idna==2.10
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.8
psutil==5.9.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
python-dateutil==2.8.2
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==2.0.0
traitlets==5.0.5
Unidecode==1.1.1
urllib3==1.25.9
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0

wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
soupsieve==2.1

wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
soupsieve==2.1
